refactor: test templates
