+++ autonomy/cli/scaffold_fsm.py
@@ -74,6 +74,17 @@ from autonomy.cli.scaffolding_templates import (
     MODEL_FILE_TEMPLATE,
     HANDLERS_FILE,
     DIALOGUES_FILE,
+    TEST_ROUNDS_FILE_HEADER,
+    TEST_ROUNDS_BASE_CLASS,
+    TEST_ROUND_CLS_TEMPLATE,
+    TEST_BEHAVIOUR_FILE_HEADER,
+    TEST_BEHAVIOUR_BASE_CLASS,
+    TEST_BEHAVIOUR_CLS_TEMPLATE,
+    TEST_PAYLOAD_FILE_HEADER,
+    TEST_PAYLOAD_CLS_TEMPLATE,
+    TEST_MODELS_FILE,
+    TEST_HANDLERS_FILE,
+    TEST_DIALOGUES_FILE,
 )
 from autonomy.constants import ABSTRACT_ROUND_ABCI_SKILL_WITH_HASH
 
@@ -186,8 +197,9 @@ class AbstractFileGenerator(ABC):
             FSMName=self.fsm_name,
             AbciApp=self.abci_app_name,
             rounds=indent(",\n".join(self.rounds), " " * 4).strip(),
+            all_rounds=indent(",\n".join(self.all_rounds), " " * 4).strip(),
+            behaviours=indent(",\n".join(self.behaviours), " " * 4).strip(),
             payloads=indent(",\n".join(self.payloads), " " * 4).strip(),
-            AbciAppCls=self.abci_app_name,
             initial_round_cls=self.dfa.default_start_state,
             initial_states=_remove_quotes(str(self.dfa.start_states)),
             transition_function=self._parse_transition_func(),
@@ -195,7 +207,7 @@ class AbstractFileGenerator(ABC):
             BaseBehaviourCls=self.abci_app_name.replace(ABCI_APP, BASE_BEHAVIOUR),
             RoundBehaviourCls=self.abci_app_name.replace(ABCI_APP, ROUND_BEHAVIOUR),
             InitialBehaviourCls=self.dfa.default_start_state.replace(ROUND, BEHAVIOUR),
-            behaviours=_remove_quotes(str(self.behaviours)),
+            round_behaviours=_remove_quotes(str(self.behaviours)),
         )
 
 
@@ -353,7 +365,7 @@ class PayloadsFileGenerator(AbstractFileGenerator):
     def _get_base_payload_section(self) -> str:
         """Get the base payload section."""
 
-        all_payloads_classes_str = [BASE_PAYLOAD_CLS.format(FSMName=self.fsm_name)]
+        all_payloads_classes_str = [BASE_PAYLOAD_CLS.format(**self.template_kwargs)]
 
         for payload_name, round_name in zip(self.payloads, self.rounds):
             tx_type = _camel_case_to_snake_case(round_name.replace(ROUND, ""))
@@ -377,7 +389,7 @@ class PayloadsFileGenerator(AbstractFileGenerator):
         return "\n".join(
             [
                 FILE_HEADER,
-                PAYLOADS_FILE.format(FSMName=self.abci_app_name),
+                PAYLOADS_FILE.format(**self.template_kwargs),
                 TRANSACTION_TYPE_SECTION.format(tx_types=tx_types),
                 self._get_base_payload_section(),
             ]
@@ -427,7 +439,7 @@ class DialoguesFileGenerator(AbstractFileGenerator):
         return "\n".join(
             [
                 FILE_HEADER,
-                DIALOGUES_FILE.format(FSMName=self.abci_app_name),
+                DIALOGUES_FILE.format(**self.template_kwargs),
             ]
         )
 
@@ -531,99 +543,6 @@ class RoundTestsFileGenerator(AbstractFileGenerator):
 
     FILENAME = "tests_" + ROUNDS_FILENAME
 
-    ROUNDS_FILE_HEADER = dedent(
-        """\
-        \"\"\"This package contains the tests for rounds of {FSMName}.\"\"\"
-
-        from typing import Any, Dict, List, Callable, Hashable
-        from dataclasses import dataclass, field
-
-        import pytest
-
-        # TODO: define and import specific payloads explicitly by name
-        from {author}.skills.{skill_name}.payloads import *
-        from {author}.skills.{skill_name}.rounds import (
-            Event,
-            SynchronizedData,
-            {non_degenerate_rounds},
-        )
-        from packages.valory.skills.abstract_round_abci.base import (
-            BaseTxPayload,
-        )
-        from packages.valory.skills.abstract_round_abci.test_tools.rounds import (
-            BaseRoundTestClass,
-            BaseOnlyKeeperSendsRoundTest,
-            BaseCollectDifferentUntilThresholdRoundTest,
-            BaseCollectSameUntilThresholdRoundTest,
-         )
-
-
-        @dataclass
-        class RoundTestCase:
-            \"\"\"RoundTestCase\"\"\"
-
-            initial_data: Dict[str, Hashable]
-            payloads: BaseTxPayload
-            final_data: Dict[str, Hashable]
-            event: Event
-            synchronized_data_attr_checks: List[Callable] = field(default_factory=list)
-
-
-        MAX_PARTICIPANTS: int = 4
-
-    """
-    )
-
-    BASE_CLASS = dedent(
-        """\
-        class Base{FSMName}RoundTestClass(BaseRoundTestClass):
-            \"\"\"Base test class for {FSMName} rounds.\"\"\"
-
-            synchronized_data: SynchronizedData
-            _synchronized_data_class = SynchronizedData
-            _event_class = Event
-
-            def run_test(self, test_case: RoundTestCase, **kwargs) -> None:
-                \"\"\"Run the test\"\"\"
-
-                self.synchronized_data.update(**test_case.initial_data)
-
-                test_round = self.round_class(
-                    synchronized_data=self.synchronized_data,
-                    consensus_params=self.consensus_params,
-                )
-
-                self._complete_run(
-                    self._test_round(
-                        test_round=test_round,
-                        round_payloads=test_case.payloads,
-                        synchronized_data_update_fn=lambda sync_data, _: sync_data.update(**test_case.final_data),
-                        synchronized_data_attr_checks=test_case.synchronized_data_attr_checks,
-                        exit_event=test_case.event,
-                        **kwargs,  # varies per BaseRoundTestClass child
-                    )
-                )
-
-            """
-    )
-
-    ROUND_CLS_TEMPLATE = dedent(
-        """\
-        class Test{RoundCls}(Base{FSMName}RoundTestClass):
-            \"\"\"Tests for {RoundCls}.\"\"\"
-
-            round_class = {RoundCls}
-
-            # TODO: provide test cases
-            @pytest.mark.parametrize("test_case, kwargs", [])
-            def test_run(self, test_case: RoundTestCase, **kwargs: Any) -> None:
-                \"\"\"Run tests.\"\"\"
-
-                self.run_test(test_case, **kwargs)
-
-    """
-    )
-
     def get_file_content(self) -> str:
         """Scaffold the 'test_rounds.py' file."""
 
@@ -643,15 +562,15 @@ class RoundTestsFileGenerator(AbstractFileGenerator):
     def _get_rounds_header_section(self) -> str:
         """Get the rounds header section."""
 
-        return self.ROUNDS_FILE_HEADER.format(**self.template_kwargs)
+        return TEST_ROUNDS_FILE_HEADER.format(**self.template_kwargs)
 
     def _get_rounds_section(self) -> str:
         """Get rounds section"""
 
-        all_round_classes_str = [self.BASE_CLASS.format(**self.template_kwargs)]
+        all_round_classes_str = [TEST_ROUNDS_BASE_CLASS.format(**self.template_kwargs)]
 
         for abci_round_name in self.dfa.states - self.dfa.final_states:
-            round_class_str = self.ROUND_CLS_TEMPLATE.format(
+            round_class_str = TEST_ROUND_CLS_TEMPLATE.format(
                 FSMName=self.fsm_name,
                 RoundCls=abci_round_name,
             )
@@ -664,107 +583,6 @@ class BehaviourTestsFileGenerator(AbstractFileGenerator):
     """File generator for 'test_behaviours.py' modules."""
 
     FILENAME = "test_" + BEHAVIOURS_FILENAME
-
-    BEHAVIOUR_FILE_HEADER = dedent(
-        """\
-        \"\"\"This package contains round behaviours of {AbciAppCls}.\"\"\"
-
-        from pathlib import Path
-        from typing import Any, Dict, Hashable, Optional, Type
-        from dataclasses import dataclass
-
-        import pytest
-
-        from packages.valory.skills.abstract_round_abci.base import AbciAppDB
-        from packages.valory.skills.abstract_round_abci.behaviours import (
-            AbstractRoundBehaviour,
-            BaseBehaviour,
-            make_degenerate_behaviour,
-        )
-        from {author}.skills.{skill_name}.behaviours import (
-            {FSMName}BaseBehaviour,
-            {behaviours},
-        )
-        from {author}.skills.{skill_name}.rounds import (
-            SynchronizedData,
-            DegenerateRound,
-            Event,
-            {AbciAppCls},
-            {rounds},
-        )
-
-        from packages.valory.skills.abstract_round_abci.test_tools.base import (
-            FSMBehaviourBaseCase,
-        )
-
-
-        @dataclass
-        class BehaviourTestCase:
-            \"\"\"BehaviourTestCase\"\"\"
-
-            initial_data: Dict[str, Hashable]
-            event: Event
-
-        """
-    )
-
-    BASE_CLASS = dedent(
-        """\
-        class Base{FSMName}Test(FSMBehaviourBaseCase):
-            \"\"\"Base test case.\"\"\"
-
-            path_to_skill = Path(__file__).parent.parent
-
-            behaviour: {FSMName}BaseBehaviour
-            behaviour_class: Type[{FSMName}BaseBehaviour]
-            next_behaviour_class: Type[{FSMName}BaseBehaviour]
-            synchronized_data: SynchronizedData
-            done_event = Event.DONE
-
-            def fast_forward(self, data: Optional[Dict[str, Any]] = None) -> None:
-                \"\"\"Fast-forward on initialization\"\"\"
-
-                data = data if data is not None else {{}}
-                self.fast_forward_to_behaviour(
-                    self.behaviour,
-                    self.behaviour_class.behaviour_id,
-                    SynchronizedData(AbciAppDB(setup_data=AbciAppDB.data_to_lists(data))),
-                )
-                assert self.behaviour.behaviour_id == self.behaviour_class.behaviour_id
-
-            def complete(self, event: Event) -> None:
-                \"\"\"Complete test\"\"\"
-
-                self.behaviour.act_wrapper()
-                self.mock_a2a_transaction()
-                self._test_done_flag_set()
-                self.end_round(done_event=event)
-                assert self.behaviour.behaviour_id == self.next_behaviour_class.behaviour_id
-
-    """
-    )
-
-    BEHAVIOUR_CLS_TEMPLATE = dedent(
-        """\
-        class Test{BehaviourCls}(Base{FSMName}Test):
-            \"\"\"Tests {BehaviourCls}\"\"\"
-
-            # TODO: set next_behaviour_class
-            behaviour_class: Type[BaseBehaviour] = {BehaviourCls}
-            next_behaviour_class: Type[BaseBehaviour] = ...
-
-            # TODO: provide test cases
-            @pytest.mark.parametrize("test_case, kwargs", [])
-            def test_run(self, test_case: BehaviourTestCase, **kwargs: Any) -> None:
-                \"\"\"Run tests.\"\"\"
-
-                self.fast_forward(test_case.initial_data)
-                # TODO: mock the necessary calls
-                # self.mock_ ...
-                self.complete(test_case.event)
-
-    """
-    )
 
     def get_file_content(self) -> str:
         """Scaffold the 'test_behaviours.py' file."""
@@ -785,28 +603,17 @@ class BehaviourTestsFileGenerator(AbstractFileGenerator):
     def _get_behaviour_header_section(self) -> str:
         """Get the rounds header section."""
 
-        return self.BEHAVIOUR_FILE_HEADER.format(
-            AbciAppCls=self.abci_app_name,
-            FSMName=self.fsm_name,
-            author=self.author,
-            skill_name=self.skill_name,
-            rounds=indent(",\n".join(self.all_rounds), " " * 4).strip(),
-            behaviours=indent(",\n".join(self.behaviours), " " * 4).strip(),
-        )
+        return TEST_BEHAVIOUR_FILE_HEADER.format(**self.template_kwargs)
 
     def _get_behaviour_section(self) -> str:
         """Get behaviour section"""
 
         all_behaviour_classes_str = [
-            self.BASE_CLASS.format(
-                FSMName=self.fsm_name,
-                author=self.author,
-                skill_name=self.skill_name,
-            )
+            TEST_BEHAVIOUR_BASE_CLASS.format(**self.template_kwargs)
         ]
 
         for behaviour_name in self.behaviours:
-            round_class_str = self.BEHAVIOUR_CLS_TEMPLATE.format(
+            round_class_str = TEST_BEHAVIOUR_CLS_TEMPLATE.format(
                 FSMName=self.fsm_name,
                 BehaviourCls=behaviour_name,
             )
@@ -819,49 +626,6 @@ class PayloadTestsFileGenerator(AbstractFileGenerator):
     """File generator for 'test_payloads.py' modules."""
 
     FILENAME = "test_" + PAYLOADS_FILENAME
-
-    PAYLOAD_FILE_HEADER = dedent(
-        """\
-        \"\"\"This package contains payload tests for the {AbciApp}.\"\"\"
-
-        from typing import Hashable
-        from dataclasses import dataclass
-
-        import pytest
-
-        from {author}.skills.{skill_name}.payloads import (
-            TransactionType,
-            Base{FSMName}Payload,
-            {payloads},
-        )
-
-
-        @dataclass
-        class PayloadTestCase:
-            \"\"\"PayloadTestCase\"\"\"
-
-            payload_cls: Base{FSMName}Payload
-            content: Hashable
-            transaction_type: TransactionType
-
-        """
-    )
-
-    PAYLOAD_CLS_TEMPLATE = dedent(
-        """\
-        # TODO: provide test cases
-        @pytest.mark.parametrize("test_case", [])
-        def test_payloads(test_case: PayloadTestCase) -> None:
-            \"\"\"Tests for {AbciApp} payloads\"\"\"
-
-            payload = test_case.payload_cls(sender="sender", content=test_case.content)
-            assert payload.sender == "sender"
-            assert getattr(payload, f"{{payload.transaction_type}}") == test_case.content
-            assert payload.transaction_type == test_case.transaction_type
-            assert payload.from_json(payload.json) == payload
-
-    """
-    )
 
     def get_file_content(self) -> str:
         """Scaffold the 'test_payloads.py' file."""
@@ -879,18 +643,12 @@ class PayloadTestsFileGenerator(AbstractFileGenerator):
     def _get_payload_header_section(self) -> str:
         """Get the rounds header section."""
 
-        return self.PAYLOAD_FILE_HEADER.format(
-            AbciApp=self.abci_app_name,
-            FSMName=self.fsm_name,
-            author=self.author,
-            skill_name=self.skill_name,
-            payloads=indent(",\n".join(self.payloads), " " * 4).strip(),
-        )
+        return TEST_PAYLOAD_FILE_HEADER.format(**self.template_kwargs)
 
     def _get_payload_section(self) -> str:
         """Get payload section"""
 
-        return self.PAYLOAD_CLS_TEMPLATE.format(AbciApp=self.abci_app_name)
+        return TEST_PAYLOAD_CLS_TEMPLATE.format(**self.template_kwargs)
 
 
 class ModelTestFileGenerator(AbstractFileGenerator):
@@ -898,32 +656,10 @@ class ModelTestFileGenerator(AbstractFileGenerator):
 
     FILENAME = "test_" + MODELS_FILENAME
 
-    MODELS_FILE = dedent(
-        """\
-        \"\"\"Test the models.py module of the {FSMName}.\"\"\"
-
-        from packages.valory.skills.abstract_round_abci.test_tools.base import DummyContext
-        from {author}.skills.{skill_name}.models import SharedState
-
-
-        class TestSharedState:
-            \"\"\"Test SharedState of {FSMName}.\"\"\"
-
-            def test_initialization(self) -> None:
-                \"\"\"Test initialization.\"\"\"
-                SharedState(name="", skill_context=DummyContext())
-
-        """
-    )
-
     def _get_models_header_section(self) -> str:
         """Get the models header section."""
 
-        return self.MODELS_FILE.format(
-            FSMName=self.fsm_name,
-            author=self.author,
-            skill_name=self.skill_name,
-        )
+        return TEST_MODELS_FILE.format(**self.template_kwargs)
 
     def get_file_content(self) -> str:
         """Get the file content."""
@@ -936,27 +672,10 @@ class HandlersTestFileGenerator(AbstractFileGenerator):
 
     FILENAME = "test_" + HANDLERS_FILENAME
 
-    HANDLERS_FILE = dedent(
-        """\
-        \"\"\"Test the handlers.py module of the {FSMName}.\"\"\"
-
-        import packages.{author}.skills.{skill_name}.handlers  # noqa
-
-
-        def test_import() -> None:
-            \"\"\"Test that the 'handlers.py' of the {FSMName} can be imported.\"\"\"
-
-        """
-    )
-
     def _get_handlers_header_section(self) -> str:
         """Get the handlers header section."""
 
-        return self.HANDLERS_FILE.format(
-            FSMName=self.fsm_name,
-            author=self.author,
-            skill_name=self.skill_name,
-        )
+        return TEST_HANDLERS_FILE.format(**self.template_kwargs)
 
     def get_file_content(self) -> str:
         """Get the file content."""
@@ -969,26 +688,10 @@ class DialoguesTestFileGenerator(AbstractFileGenerator):
 
     FILENAME = "test_" + DIALOGUES_FILENAME
 
-    DIALOGUES_FILE = dedent(
-        """\
-        \"\"\"Test the dialogues.py module of the {FSMName}.\"\"\"
-
-        import packages.{author}.skills.{skill_name}.dialogues  # noqa
-
-
-        def test_import() -> None:
-            \"\"\"Test that the 'dialogues.py' of the {FSMName} can be imported.\"\"\"
-        """
-    )
-
     def _get_dialogues_header_section(self) -> str:
         """Get the dialogues header section."""
 
-        return self.DIALOGUES_FILE.format(
-            FSMName=self.fsm_name,
-            author=self.author,
-            skill_name=self.skill_name,
-        )
+        return TEST_DIALOGUES_FILE.format(**self.template_kwargs)
 
     def get_file_content(self) -> str:
         """Get the file content."""
