+++ autonomy/cli/scaffolding_templates/__init__.py
@@ -96,8 +96,8 @@ class {RoundCls}({ABCRoundCls}):
 """
 
 ABCI_APP_CLS_TEMPLATE = """\
-class {AbciAppCls}(AbciApp[Event]):
-    \"\"\"{AbciAppCls}\"\"\"
+class {AbciApp}(AbciApp[Event]):
+    \"\"\"{AbciApp}\"\"\"
 
     initial_round_cls: AppState = {initial_round_cls}
     initial_states: Set[AppState] = {initial_states}
@@ -166,8 +166,8 @@ class {RoundBehaviourCls}(AbstractRoundBehaviour):
     \"\"\"{RoundBehaviourCls}\"\"\"
 
     initial_behaviour_cls = {InitialBehaviourCls}
-    abci_app_cls = {AbciAppCls}  # type: ignore
-    behaviours: Set[Type[BaseBehaviour]] = {behaviours}
+    abci_app_cls = {AbciApp}  # type: ignore
+    behaviours: Set[Type[BaseBehaviour]] = {round_behaviours}
 """
 
 
@@ -349,4 +349,278 @@ ContactApiDialogues = BaseContractApiDialogues
 
 TendermintDialogue = BaseTendermintDialogue
 TendermintDialogues = BaseTendermintDialogues
+"""
+
+
+# Tests
+
+
+# Test rounds
+TEST_ROUNDS_FILE_HEADER = """\
+\"\"\"This package contains the tests for rounds of {FSMName}.\"\"\"
+
+from typing import Any, Dict, List, Callable, Hashable
+from dataclasses import dataclass, field
+
+import pytest
+
+# TODO: define and import specific payloads explicitly by name
+from {author}.skills.{skill_name}.payloads import *
+from {author}.skills.{skill_name}.rounds import (
+    Event,
+    SynchronizedData,
+    {rounds},
+)
+from packages.valory.skills.abstract_round_abci.base import (
+    BaseTxPayload,
+)
+from packages.valory.skills.abstract_round_abci.test_tools.rounds import (
+    BaseRoundTestClass,
+    BaseOnlyKeeperSendsRoundTest,
+    BaseCollectDifferentUntilThresholdRoundTest,
+    BaseCollectSameUntilThresholdRoundTest,
+ )
+
+
+@dataclass
+class RoundTestCase:
+    \"\"\"RoundTestCase\"\"\"
+
+    initial_data: Dict[str, Hashable]
+    payloads: BaseTxPayload
+    final_data: Dict[str, Hashable]
+    event: Event
+    synchronized_data_attr_checks: List[Callable] = field(default_factory=list)
+
+
+MAX_PARTICIPANTS: int = 4
+
+"""
+
+
+TEST_ROUNDS_BASE_CLASS = """\
+class Base{FSMName}RoundTestClass(BaseRoundTestClass):
+    \"\"\"Base test class for {FSMName} rounds.\"\"\"
+
+    synchronized_data: SynchronizedData
+    _synchronized_data_class = SynchronizedData
+    _event_class = Event
+
+    def run_test(self, test_case: RoundTestCase, **kwargs) -> None:
+        \"\"\"Run the test\"\"\"
+
+        self.synchronized_data.update(**test_case.initial_data)
+
+        test_round = self.round_class(
+            synchronized_data=self.synchronized_data,
+            consensus_params=self.consensus_params,
+        )
+
+        self._complete_run(
+            self._test_round(
+                test_round=test_round,
+                round_payloads=test_case.payloads,
+                synchronized_data_update_fn=lambda sync_data, _: sync_data.update(**test_case.final_data),
+                synchronized_data_attr_checks=test_case.synchronized_data_attr_checks,
+                exit_event=test_case.event,
+                **kwargs,  # varies per BaseRoundTestClass child
+            )
+        )
+
+"""
+
+TEST_ROUND_CLS_TEMPLATE = """\
+class Test{RoundCls}(Base{FSMName}RoundTestClass):
+    \"\"\"Tests for {RoundCls}.\"\"\"
+
+    round_class = {RoundCls}
+
+    # TODO: provide test cases
+    @pytest.mark.parametrize("test_case, kwargs", [])
+    def test_run(self, test_case: RoundTestCase, **kwargs: Any) -> None:
+        \"\"\"Run tests.\"\"\"
+
+        self.run_test(test_case, **kwargs)
+
+"""
+
+
+# Test behaviours
+TEST_BEHAVIOUR_FILE_HEADER = """\
+\"\"\"This package contains round behaviours of {AbciApp}.\"\"\"
+
+from pathlib import Path
+from typing import Any, Dict, Hashable, Optional, Type
+from dataclasses import dataclass
+
+import pytest
+
+from packages.valory.skills.abstract_round_abci.base import AbciAppDB
+from packages.valory.skills.abstract_round_abci.behaviours import (
+    AbstractRoundBehaviour,
+    BaseBehaviour,
+    make_degenerate_behaviour,
+)
+from {author}.skills.{skill_name}.behaviours import (
+    {FSMName}BaseBehaviour,
+    {behaviours},
+)
+from {author}.skills.{skill_name}.rounds import (
+    SynchronizedData,
+    DegenerateRound,
+    Event,
+    {AbciApp},
+    {all_rounds},
+)
+
+from packages.valory.skills.abstract_round_abci.test_tools.base import (
+    FSMBehaviourBaseCase,
+)
+
+
+@dataclass
+class BehaviourTestCase:
+    \"\"\"BehaviourTestCase\"\"\"
+
+    initial_data: Dict[str, Hashable]
+    event: Event
+
+"""
+
+
+TEST_BEHAVIOUR_BASE_CLASS = """\
+class Base{FSMName}Test(FSMBehaviourBaseCase):
+    \"\"\"Base test case.\"\"\"
+
+    path_to_skill = Path(__file__).parent.parent
+
+    behaviour: {FSMName}BaseBehaviour
+    behaviour_class: Type[{FSMName}BaseBehaviour]
+    next_behaviour_class: Type[{FSMName}BaseBehaviour]
+    synchronized_data: SynchronizedData
+    done_event = Event.DONE
+
+    def fast_forward(self, data: Optional[Dict[str, Any]] = None) -> None:
+        \"\"\"Fast-forward on initialization\"\"\"
+
+        data = data if data is not None else {{}}
+        self.fast_forward_to_behaviour(
+            self.behaviour,
+            self.behaviour_class.behaviour_id,
+            SynchronizedData(AbciAppDB(setup_data=AbciAppDB.data_to_lists(data))),
+        )
+        assert self.behaviour.behaviour_id == self.behaviour_class.behaviour_id
+
+    def complete(self, event: Event) -> None:
+        \"\"\"Complete test\"\"\"
+
+        self.behaviour.act_wrapper()
+        self.mock_a2a_transaction()
+        self._test_done_flag_set()
+        self.end_round(done_event=event)
+        assert self.behaviour.behaviour_id == self.next_behaviour_class.behaviour_id
+
+"""
+
+
+TEST_BEHAVIOUR_CLS_TEMPLATE = """\
+class Test{BehaviourCls}(Base{FSMName}Test):
+    \"\"\"Tests {BehaviourCls}\"\"\"
+
+    # TODO: set next_behaviour_class
+    behaviour_class: Type[BaseBehaviour] = {BehaviourCls}
+    next_behaviour_class: Type[BaseBehaviour] = ...
+
+    # TODO: provide test cases
+    @pytest.mark.parametrize("test_case, kwargs", [])
+    def test_run(self, test_case: BehaviourTestCase, **kwargs: Any) -> None:
+        \"\"\"Run tests.\"\"\"
+
+        self.fast_forward(test_case.initial_data)
+        # TODO: mock the necessary calls
+        # self.mock_ ...
+        self.complete(test_case.event)
+
+"""
+
+
+# Test payloads
+TEST_PAYLOAD_FILE_HEADER = """\
+\"\"\"This package contains payload tests for the {AbciApp}.\"\"\"
+
+from typing import Hashable
+from dataclasses import dataclass
+
+import pytest
+
+from {author}.skills.{skill_name}.payloads import (
+    TransactionType,
+    Base{FSMName}Payload,
+    {payloads},
+)
+
+
+@dataclass
+class PayloadTestCase:
+    \"\"\"PayloadTestCase\"\"\"
+
+    payload_cls: Base{FSMName}Payload
+    content: Hashable
+    transaction_type: TransactionType
+
+"""
+
+
+TEST_PAYLOAD_CLS_TEMPLATE = """\
+# TODO: provide test cases
+@pytest.mark.parametrize("test_case", [])
+def test_payloads(test_case: PayloadTestCase) -> None:
+    \"\"\"Tests for {AbciApp} payloads\"\"\"
+
+    payload = test_case.payload_cls(sender="sender", content=test_case.content)
+    assert payload.sender == "sender"
+    assert getattr(payload, f"{{payload.transaction_type}}") == test_case.content
+    assert payload.transaction_type == test_case.transaction_type
+    assert payload.from_json(payload.json) == payload
+
+"""
+
+
+TEST_MODELS_FILE = """\
+\"\"\"Test the models.py module of the {FSMName}.\"\"\"
+
+from packages.valory.skills.abstract_round_abci.test_tools.base import DummyContext
+from {author}.skills.{skill_name}.models import SharedState
+
+
+class TestSharedState:
+    \"\"\"Test SharedState of {FSMName}.\"\"\"
+
+    def test_initialization(self) -> None:
+        \"\"\"Test initialization.\"\"\"
+        SharedState(name="", skill_context=DummyContext())
+
+"""
+
+
+TEST_HANDLERS_FILE = """\
+\"\"\"Test the handlers.py module of the {FSMName}.\"\"\"
+
+import packages.{author}.skills.{skill_name}.handlers  # noqa
+
+
+def test_import() -> None:
+    \"\"\"Test that the 'handlers.py' of the {FSMName} can be imported.\"\"\"
+
+"""
+
+
+TEST_DIALOGUES_FILE = """\
+\"\"\"Test the dialogues.py module of the {FSMName}.\"\"\"
+
+import packages.{author}.skills.{skill_name}.dialogues  # noqa
+
+
+def test_import() -> None:
+    \"\"\"Test that the 'dialogues.py' of the {FSMName} can be imported.\"\"\"
 """
